+++ src/models/index.ts
@@ -955,9 +955,9 @@ export interface Annotation {
   /** x/y coordinates for the rectangular bounding box containing the data */
   rectangle: Rectangle | null;
   /** x/y coordinates for the rectangles containing the data. An annotation can be contained within multiple rectangles. */
-  rectangles: Rectangle[] | null;
+  rectangles: Rectangle[];
   /** Unique identifier for the document */
-  document?: string;
+  document: string;
   /** The page number within the document, starting from 0. */
   pageIndex: number | null;
   /** Raw data extracted from the before any post-processing */
@@ -978,6 +978,8 @@ export interface Annotation {
   dataPoint: string;
   /** The different data types of annotations */
   contentType: AnnotationContentType;
+  /** The parent annotation's ID */
+  parent?: number;
 }
 
 export interface Rectangle {
@@ -1912,7 +1914,7 @@ export interface Get200ApplicationJsonPropertiesItemsItem {
 
 export interface Paths1TvfqeiV3IndexPostResponses201ContentApplicationJsonSchema {
   name?: string;
-  documentType?: Enum23;
+  documentType?: Enum22;
 }
 
 export interface PathsO7SnenV3IndexNameDocumentsGetResponses200ContentApplicationJsonSchema {
@@ -2945,24 +2947,6 @@ export enum KnownDateRange {
  */
 export type DateRange = string;
 
-/** Known values of {@link Get8ItemsItem} that the service accepts. */
-export enum KnownGet8ItemsItem {
-  FileName = "file_name",
-  Extractor = "extractor",
-  CreatedDt = "created_dt"
-}
-
-/**
- * Defines values for Get8ItemsItem. \
- * {@link KnownGet8ItemsItem} can be used interchangeably with Get8ItemsItem,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **file_name** \
- * **extractor** \
- * **created_dt**
- */
-export type Get8ItemsItem = string;
-
 /** Known values of {@link ResumeDataLanguagesItem} that the service accepts. */
 export enum KnownResumeDataLanguagesItem {
   Ainu = "Ainu",
@@ -3527,21 +3511,21 @@ export enum KnownSearchParametersCustomDataFilterType {
  */
 export type SearchParametersCustomDataFilterType = string;
 
-/** Known values of {@link Enum20} that the service accepts. */
-export enum KnownEnum20 {
+/** Known values of {@link Enum19} that the service accepts. */
+export enum KnownEnum19 {
   Resumes = "resumes",
   JobDescriptions = "job_descriptions"
 }
 
 /**
- * Defines values for Enum20. \
- * {@link KnownEnum20} can be used interchangeably with Enum20,
+ * Defines values for Enum19. \
+ * {@link KnownEnum19} can be used interchangeably with Enum19,
  *  this enum contains the known values that the service supports.
  * ### Known values supported by the service
  * **resumes** \
  * **job_descriptions**
  */
-export type Enum20 = string;
+export type Enum19 = string;
 
 /** Known values of {@link GetResponses200ContentApplicationJsonSchemaResultsItemDocumentType} that the service accepts. */
 export enum KnownGetResponses200ContentApplicationJsonSchemaResultsItemDocumentType {
@@ -3575,21 +3559,21 @@ export enum KnownPostContentSchemaDocumentType {
  */
 export type PostContentSchemaDocumentType = string;
 
-/** Known values of {@link Enum23} that the service accepts. */
-export enum KnownEnum23 {
+/** Known values of {@link Enum22} that the service accepts. */
+export enum KnownEnum22 {
   Resumes = "resumes",
   JobDescriptions = "job_descriptions"
 }
 
 /**
- * Defines values for Enum23. \
- * {@link KnownEnum23} can be used interchangeably with Enum23,
+ * Defines values for Enum22. \
+ * {@link KnownEnum22} can be used interchangeably with Enum22,
  *  this enum contains the known values that the service supports.
  * ### Known values supported by the service
  * **resumes** \
  * **job_descriptions**
  */
-export type Enum23 = string;
+export type Enum22 = string;
 
 /** Known values of {@link ResumeSkillSourcesItemSection} that the service accepts. */
 export enum KnownResumeSkillSourcesItemSection {
@@ -3819,8 +3803,8 @@ export interface AffindaAPIGetAllDocumentsOptionalParams
   createdDt?: DateRange;
   /** Partial, case-insensitive match with file name or tag name. */
   search?: string;
-  /** Sort the result set. A "-" at the beginning denotes DESC sort, e.g. -created_dt. Sort by multiple fields is supported. */
-  ordering?: Get8ItemsItem[];
+  /** Sort the result set. A "-" at the beginning denotes DESC sort, e.g. -created_dt. Sort by multiple fields is supported. Supported values include: 'file_name', 'extractor', 'created_dt', 'validated_dt', 'archived_dt' and 'parsed__<dataPointSlug>'. */
+  ordering?: string[];
   /** By default, this endpoint returns only the meta data of the documents. Set this to `true` will return a summary of the data that was parsed. If you want to retrieve the full set of data for a document, use the `GET /documents/{identifier}` endpoint. */
   includeData?: boolean;
   /** Exclude some documents from the result. */
@@ -4434,7 +4418,7 @@ export interface AffindaAPIGetAllIndexesOptionalParams
   /** The numbers of results to return. */
   limit?: number;
   /** Filter indices by a document type */
-  documentType?: Enum20;
+  documentType?: Enum19;
 }
 
 /** Contains response data for the getAllIndexes operation. */